requirements/requirements.txt
numpy
pyyaml
xtgeo
fmu-dataio>=1.0.0
xtgeoviz
